refactor: compress the packages index
We refactor the implementation of the mapping that allows us to convert
a file path to the corresponding distribution. Instead of mapping each
individual file path to the containing distribution, we extract the root
modules and map those to the distribution instead. This leads to a
significant compression of the in-memory mapping. Part of the saved
memory can thus be used to cache the result of the conversions to avoid
repeating unnecessary mapping work.