fix(llmobs): ensure propagation of ml_app/session_id fields (#9028)
This PR introuduces a workaround to LLMObs ml_app/session_id
propagation.

The expected behavior is that if a user does not explicitly set
`ml_app/session_id` when starting a LLMObs span, we should check the
nearest LLMObs ancestor span to propagate their `ml_app/session_id`
downwards. Currently, this is broken since we don't actually set the
corresponding tag on the downstream spans (which is what we check for
when we process that span's children/descendants.)

The workaround right now is to set a private tag on each span each time
we process it, to ensure its descendant spans will be able to look it
up. The con is that we can't actually remove it due to the order of span
processing in a trace - the oldest spans get processed first, meaning if
we delete the span then its children can't access it. Luckily this is a
private tag and doesn't get displayed on APM side.

The correct fix is to make sure all of this propagation and checks
happen at span start time (since oldest spans start first, we can
propagate the tags then), and delete the temporary tags at processing
time. A complication is that this requires refactoring the
OpenAI/LangChain/Bedrock integrations and corresponding LLM integration
classes to set those temporary tags at span start time, and that's a lot
of work. This will be done in a subsequent PR, this PR is to provide an
immediate workaround.

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)