fix(bootstrap): do not load new modules after module cleanup [backport #7649 to 1.20] (#7748)
Backport of #7649 to 1.20

We refactor the logic of the sitecustomize bootstrap script to make it
harder to introduce new initialisation logic that can break some of the
boilerplate bootstrapping logic, such as the module cleanup mechanism
for the support of frameworks like gevent. We move all the logic into a
dedicated module, preload, and expose post_preload hooks in case that
some code needs to be run after the module cleanup. However, the problem
is now slightly shifted towards these hooks. Care must be taken not to
register any hooks that import modules that are not already imported
during the execution of the preload script.

## Testing Strategy

The test that was supposed to catch the presence of the `threading`
module after initialisation when cleaning up the imported modules failed
because RC was disabled globally in tests with
`DD_REMOTE_CONFIGURATION_ENABLED=1` in the `riotfile.py` script. We have
modified the test to ensure that RC is enabled.

## Further Details

Because the changes in this PR include a refactor, we summarise the part
of the old coding that actually caused the issue in the first place. The
`sitecustomize.py` source included the following coding

```python
    # Only the import of the original sitecustomize.py is allowed after this
    # point.

    if sys.version_info >= (3, 7) and config._otel_enabled:
        from opentelemetry.trace import set_tracer_provider

        from ddtrace.opentelemetry import TracerProvider

        set_tracer_provider(TracerProvider())

    ...

    if config._remote_config_enabled:
        from ddtrace.internal.remoteconfig.worker import remoteconfig_poller

        remoteconfig_poller.enable()

    if asm_config._asm_enabled or config._remote_config_enabled:
        from ddtrace.appsec._remoteconfiguration import enable_appsec_rc

        enable_appsec_rc()
```

The first `if` block introduces two problems:
- it force-loads the `opentelemetry` module
- it imports more modules from `ddtrace` _after_ the comment that
instructs not to add imports beyond that point

Note that the current implementation of the `opentelemetry` intergration
also force-loads the `opentelemetry` module when `TracerProvider` is
imported. Therefore, the whole logic should be registered as a module
import hook.

The other `if` blocks further down in the `sitecustomize.py` source also
import from `ddtrace`. In fact, they import features that require
threading support, which ultimately forces the gevent-crucial
`threading` module to be re-loaded too early after the loaded module
clean-up.

## Checklist

- [x] Change(s) are motivated and described in the PR description.
- [x] Testing strategy is described if automated tests are not included
in the PR.
- [x] Risk is outlined (performance impact, potential for breakage,
maintainability, etc).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed. If no release note is required, add label
`changelog/no-changelog`.
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/)).
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist

- [x] Title is accurate.
- [x] No unnecessary changes are introduced.
- [x] Description motivates each change.
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes unless absolutely necessary.
- [x] Testing strategy adequately addresses listed risk(s).
- [x] Change is maintainable (easy to change, telemetry, documentation).
- [x] Release note makes sense to a user of the library.
- [x] Reviewer has explicitly acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment.
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
- [x] If this PR touches code that signs or publishes builds or
packages, or handles credentials of any kind, I've requested a review
from `@DataDog/security-design-and-guidance`.
- [x] This PR doesn't touch any of that.

---------

Co-authored-by: Tahir H. Butt <tahir.butt@datadoghq.com>
Co-authored-by: Alberto Vara <alberto.vara@datadoghq.com>
Co-authored-by: Emmett Butler <723615+emmettbutler@users.noreply.github.com>
Co-authored-by: Emmett Butler <emmett.butler321@gmail.com>