chore(ci): add OCI packaging (#8932)
We are shifting over to using OCI packaging rather than system
packaging. Unlike the existing deb/rpm builds we package only the
relevant arch binaries in the package to reduce the final installation
size on the system. Once validated and released we will remove the
similar deb/rpm packaging.

This packaging has been tested in
https://github.com/DataDog/auto_inject/pull/185 where we pull in v2.8.2
of the library and inject it into a basic Python app and validate that
traces are submitted.

The job will run as part of the existing deb/rpm jobs done as part of
the release process. No new manual steps are introduced to the release
to build and deploy these images.

(cherry picked from commit fcfe10c7039919a649ccc13a3dc3a53c510b55d9)