chore(internal): bytecode wrapping context
We introduce a new mechanism of wrapping functions via a special context
manager that is capable of capturing return values as well. The goal is
to allow observability into the called functions, to have access to
local variables on exit. This approach has the extra benefit of not
introducing any extra frames in the call stack of the wrapped function.