fix(ci_visibility): use the default tracer when enabling CI Visibility (#9016)
This switches back to using the default tracer when using CI Visiblity,
and fixes an issue where non-test traces were getting lost due to:
* going through a non-existent local agent when using
`DD_CIVISIBILITY_AGENTLESS_ENABLED=true`
* being in a different tracer object which broke trace IDs

The historical change behind switching to a separate tracer was so that
we would be able to use CI Visibility in this repo, which was causing CI
Visibility traces to get mixed up with other traces. Since that would
remain a problem the behavior to use a separate tracer with a separate
context provider, via the use of
`_DD_CIVISIBILITY_USE_CI_CONTEXT_PROVIDER`, is maintained.

This also tweaks the `TraceCiVisibilityFilter` to no longer force the
`service` of spans to match that of the original under which the tracer
was started, which allows for test-level traces with non-test spans (eg:
with distributed tracing or `--ddtrace-patch-all` enabled) to properly
reflect the right services.

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

(cherry picked from commit 6afcedbad8e7bc18c7b618f3660945636394898f)