feat(tracer): add `TracerFlareSubscriber` to enable the tracer flare (#9150)
## Overview
This PR adds a remote config subscriber for the tracer flare. The tracer
flare logic was already implemented in a previous PR, but was not wired
up to actually react to the RC products yet. This PR is the last piece
to enable the tracer flare for dd-trace-py.

A tracer flare is a tar file containing tracer logs from (approximately)
the last 5 minutes, as well as a JSON file of the current tracer
configurations. This pair of files will be generated per tracer instance
that is connected to the agent, and the tar containing all of these
files will be sent to a Zendesk ticket. Users can trigger the tracer
flare the same way they trigger agent flares. See Agent Flare
documentation
[here](https://docs.datadoghq.com/agent/troubleshooting/send_a_flare/?tab=agentv6v7).

For details on the flare implementation, see
https://github.com/DataDog/dd-trace-py/pull/8961 and
https://github.com/DataDog/dd-trace-py/pull/8969.

## Risks
### `AGENT_CONFIG` doesn't get cleared very frequently
Something I noticed when doing some E2E testing is that if you try to do
consecutive tracer flare requests, this won't work because the
`AGENT_CONFIG` is still retaining the state from the previous request.
This means that there isn't an update/publish event that can get picked
up on our end, so we can't trigger another flare for some amount of time
(not sure what this duration is exactly). The current implementation
depends on the publish event, so trying to trigger consecutive tracer
flare requests will not work until the state gets cleared.

### `AGENT_CONFIG` and `AGENT_TASK` are not exclusive to tracer flare
use
Currently, the tracer is listening for changes to the `AGENT_CONFIG` and
`AGENT_TASK` remote config products. This is originally intended for the
**agent** flare, not the tracer flare, but at this time we are
piggy-backing on this signal. For this reason, it's been flagged by
other tracer teams that the format/contents of the products may not be
guaranteed. In the case that we start to notice flares not being
triggered/generated as expected, this may be a code fix to check for.
The current expectation for the products is:

`AGENT_CONFIG`
```json
{
   "metadata":[
      {
         "id":"flare-log-level.<log-level>",
         "product_name":"AGENT_CONFIG",
         "sha256_hash":"xxx",
         "length":63,
         "tuf_version":3,
         "apply_state":2,
         "apply_error":"None"
      }
   ],
   "config":[
      {
         "config":{
            "log_level":"<log-level>"
         },
         "name":"flare-log-level.<log-level>"
      }
   ],
   "shared_data_counter":2
}
```

`AGENT_TASK`
```json
{
   "metadata":[
      {
         "id":"id1",
         "product_name":"AGENT_TASK",
         "sha256_hash":"xxx",
         "length":139,
         "tuf_version":4,
         "apply_state":2,
         "apply_error":"None"
      }
   ],
   "config":[
      false,
      {
         "args":{
            "case_id":"111",
            "hostname":"myhostname",
            "user_handle":"user.name@datadoghq.com"
         },
         "task_type":"tracer_flare",
         "uuid":"yyyyyy"
      }
   ],
   "shared_data_counter":5
}

```

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Co-authored-by: Brett Langdon <brett.langdon@datadoghq.com>