fix(asm): initializer rework and fix for segmentation faults (#9284)
Fix several segmentation faults (3 confirmed, probably more).

This is done by a small rework of the Initializer which includes:

- Moving `active_map_addreses` from `unordered_set` to `unordered_map`
so we can iterate and change it without making a copy (making a copy was
actually causing one of the segmentation faults because we were setting
the `tx_map` pointer to `nullptr`, but on the copy, not the original, so
`get_tainting_map` would return `not nullptr` but still garbage since it
was deleted).

- Convert the `TaintRangeMapType` pointer to a `shared_ptr`. This fixed
another segfault and makes the codebase more robust.

- Make `ThreadContextCache` not use a dangerously magical number-pointer
but just a reference to the active `tx_map` so we only need to check if
it's `nullptr` to see if we are in context (which is what functions did
anyway).

- Removed `destroy_context`. Not needed anymore, `reset_context` will do
the same in a safer way.

- [X] Change(s) are motivated and described in the PR description
- [X] Testing strategy is described if automated tests are not included
in the PR
- [X] Risks are described (performance impact, potential for breakage,
maintainability)
- [X] Change is maintainable (easy to change, telemetry, documentation)
- [X] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [X] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [X] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [X] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

---------

Signed-off-by: Juanjo Alvarez <juanjo.alvarezmartinez@datadoghq.com>
Co-authored-by: Federico Mon <federico.mon@datadoghq.com>
(cherry picked from commit de69597a925888ab00b99e8d84aa948c18d8bbf4)