fix(tracer): ensure all metadata and configs are checked for flare + correct config file (#9322)
## Overview
This PR fixes two bugs with the tracer flare.

First, the RC poller behavior for the tracer flare. Prior to this PR, it
would assume that the first product when polling the RC backend would be
the target config/metadata we need. However, this is not a safe
assumption as multiple products could be updated and polled at once.
This, in addition to the fact that currently the tracer flare shares the
`AGENT_CONFIG` and `AGENT_TASK` product with the **agent** flare, means
that we need to check all products to ensure we don't accidentally skip
a tracer flare product.

Second, we were passing in the RC config, not the actual ddtrace config
when generating the config file for the flare. This is fixed, and the
correct config is being generated now.

## Checklist

- [x] Change(s) are motivated and described in the PR description
- [x] Testing strategy is described if automated tests are not included
in the PR
- [x] Risks are described (performance impact, potential for breakage,
maintainability)
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] [Library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
are followed or label `changelog/no-changelog` is set
- [x] Documentation is included (in-code, generated user docs, [public
corp docs](https://github.com/DataDog/documentation/))
- [x] Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))
- [x] If this PR changes the public interface, I've notified
`@DataDog/apm-tees`.

## Reviewer Checklist

- [x] Title is accurate
- [x] All changes are related to the pull request's stated goal
- [x] Description motivates each change
- [x] Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- [x] Testing strategy adequately addresses listed risks
- [x] Change is maintainable (easy to change, telemetry, documentation)
- [x] Release note makes sense to a user of the library
- [x] Author has acknowledged and discussed the performance implications
of this PR as reported in the benchmarks PR comment
- [x] Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)