chore(di): report stack trace and tracebacks correctly
We fix the way stacktraces and tracebacks are reported in snapshot logs.
We ensure that the line number information for the top frame from
function probes is adjusted with the information from the traceback
object to ensure precise location information. We also make sure to
report just the traceback with uncaught exceptions.