chore: restore module loader after import
We restore the original loader after a module has been imported. We only
need our custom chained loader while the module is being imported to
allow the import hooks to be triggered at the right time. There is no
need for the custom loader once the module is fully imported.