chore(tracer): span exception core event
We add the `span.exception` core event to the `Span.set_exc_info`
method to notify that an exception is being attached to a span.
Other products needing to be notified of this event can subscribe
to it to get details about the span and the exception being
attached to it.